RoomSleepTracker/app/src/main/java/com/samsad/roomsleeptracker/database/SleepDatabase.kt
package com.samsad.roomsleeptracker.database

import android.content.Context
import androidx.room.Database
import androidx.room.Room
import androidx.room.RoomDatabase

@Database(entities = [SleepNight::class], version = 1, exportSchema = false)
abstract class SleepDatabase : RoomDatabase() {

    abstract val sleepDatabaseDao: SleepDatabaseDao

    companion object {
        //Volatile keyword will help to get values from memory and block caching
        @Volatile
        private var INSTANCE: SleepDatabase? = null

        fun getInstance(context: Context): SleepDatabase {
            //Multiple threads can ask for database instance at same time.
            //Wrapping in synchronised only one thread of execution at a time can
            //enter this block of code which makes sure db gets initialise only once
            synchronized(this) {
                var instance = INSTANCE

                if (instance == null) {
                    instance = Room.databaseBuilder(
                        context.applicationContext,
                        SleepDatabase::class.java,
                        "sleep_history_database"
                    ).fallbackToDestructiveMigration()
                        .build()
                    INSTANCE = instance
                }

                return instance
            }

        }
    }

}