RoomSleepTracker/app/src/androidTest/java/com/samsad/roomsleeptracker/SleepDatabaseTest.kt
package com.samsad.roomsleeptracker

import androidx.room.Room
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import com.samsad.roomsleeptracker.database.SleepDatabase
import com.samsad.roomsleeptracker.database.SleepDatabaseDao
import com.samsad.roomsleeptracker.database.SleepNight
import junit.framework.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException

@RunWith(AndroidJUnit4::class)
class SleepDatabaseTest {
    private lateinit var sleepDao: SleepDatabaseDao
    private lateinit var db: SleepDatabase

    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        //Using in memory database becuase the information stored here disappears when the
        //process is killed
        db = Room.inMemoryDatabaseBuilder(context, SleepDatabase::class.java)
            //Allowing main thread queries,just for testing
            .allowMainThreadQueries()
            .build()
        sleepDao = db.sleepDatabaseDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGetNights() {
        val night = SleepNight()
        sleepDao.insert(night)
        val tonight = sleepDao.getTonight()
        assertEquals(tonight?.sleepQuality, -1)
    }


}